wip: migrate cwlog api
